add download logic and test
