+++ Dockerfile
@@ -2,7 +2,7 @@ FROM golang:latest
 
 WORKDIR /app
 COPY go.mod .
-# COPY go.sum .
+COPY go.sum .
 RUN mkdir upload
 RUN go mod download
 COPY . .
