+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/tttol/mos3
 
 go 1.21.3
+
+require github.com/go-playground/assert/v2 v2.2.0
